Dockerfile: fix simplicity extension symlink issue under rootless podman
Using the build workflow with rootless podman containers and volumes:

 podman \
         run --rm -it \
         --privileged \
         -v $(PWD)/build:/build:Z \
         silabs-firmware-builder \
         bash -c " \
                 cd /build && \
                 ./tools/build_firmware.sh test.yml \
         "

results into following permssions issue:

 The sdk /simplicity_sdk_2024.6.2/ ( com.silabs.sdk.stack.sisdk:2024.6.2._-620023087 )  is now trusted.
 ln: failed to create symbolic link '/simplicity_sdk_2024.6.2/extension': Permission denied

This is happening due to the user/group mapping between container and
the host and currently the simplicity_sdk_2024 directory is owned as
root, thus builder user won't be able to create an extension symlink,
resulting in this failure.

So lets fix it by chown-ing the simplicity_sdk_2024 for builder.

References: https://www.redhat.com/en/blog/debug-rootless-podman-mounted-volumes
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>